requirements.txt
requests==2.31.0
Pillow==10.0.0
RPi.GPIO==0.7.1
spidev==3.6
waveshare-epd==0.0.0